requirements.txt
opencv-python-headless
mtcnn
Pillow
numpy
matplotlib
streamlit
